requirements.txt
# Core dependencies
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.0
typer[all]==0.9.0
rich==13.7.0

# Development dependencies
pytest==8.0.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
